[RENAME]/#9 PackingList를 List로 변경
